style(tests/audio_file,correlation): silence -Werror findings (Mi-18)
test_audio_file.cpp: mark create_test_wav [[maybe_unused]] (test
helper that's no longer wired up; flagged in m-i-18-followups.md).

test_correlation.cpp: spot-cast M_PI/index/seed values to float in
sine-generation expressions to avoid double-promotion, cast int64_t
true_lag to size_t at subscript sites, and cast rms to double in the
WithinAbs matcher to silence Catch2's float->double promotion in
catch_matchers_impl.hpp.

+++ tests/test_correlation.cpp
@@ -8,7 +8,7 @@ TEST_CASE("Cross-correlation finds correct lag for shifted signal", "[correlatio
     // Create a simple signal
     std::vector<float> original(1000);
     for (size_t i = 0; i < original.size(); ++i) {
-        original[i] = std::sin(2.0f * M_PI * 10.0f * i / 1000.0f);
+        original[i] = std::sin(2.0f * static_cast<float>(M_PI) * 10.0f * static_cast<float>(i) / 1000.0f);
     }
     
     // Shift by 100 samples
@@ -29,23 +29,23 @@ TEST_CASE("RMS normalization produces unit energy", "[correlation]") {
 
     float sum_sq = 0.0f;
     for (float s : samples) sum_sq += s * s;
-    float rms = std::sqrt(sum_sq / samples.size());
+    float rms = std::sqrt(sum_sq / static_cast<float>(samples.size()));
 
-    REQUIRE_THAT(rms, Catch::Matchers::WithinAbs(1.0, 0.01));
+    REQUIRE_THAT(static_cast<double>(rms), Catch::Matchers::WithinAbs(1.0, 0.01));
 }
 
 TEST_CASE("FFT cross-correlation finds correct lag for shifted signal", "[correlation][fft]") {
     // Reference: 1000-sample sinusoid with some DC added (tests mean-centering)
     std::vector<float> reference(1000);
     for (size_t i = 0; i < reference.size(); ++i) {
-        reference[i] = 0.1f + std::sin(2.0f * M_PI * 7.0f * i / 1000.0f);
+        reference[i] = 0.1f + std::sin(2.0f * static_cast<float>(M_PI) * 7.0f * static_cast<float>(i) / 1000.0f);
     }
 
     // Target: the reference placed at a known offset inside a longer buffer
     const int64_t true_lag = 347;
     std::vector<float> target(2500, 0.0f);
     for (size_t i = 0; i < reference.size(); ++i) {
-        target[true_lag + i] = reference[i];
+        target[static_cast<std::size_t>(true_lag) + i] = reference[i];
     }
 
     auto result = mwaac::cross_correlate_fft(reference, target);
@@ -57,8 +57,8 @@ TEST_CASE("FFT cross-correlation with noise still locks on target", "[correlatio
     // Reference: deterministic "distinctive" signal
     std::vector<float> reference(2000);
     for (size_t i = 0; i < reference.size(); ++i) {
-        reference[i] = std::sin(2.0f * M_PI * 13.0f * i / 200.0f)
-                     + 0.5f * std::sin(2.0f * M_PI * 3.0f * i / 200.0f);
+        reference[i] = std::sin(2.0f * static_cast<float>(M_PI) * 13.0f * static_cast<float>(i) / 200.0f)
+                     + 0.5f * std::sin(2.0f * static_cast<float>(M_PI) * 3.0f * static_cast<float>(i) / 200.0f);
     }
 
     // Target: reference placed at a known lag, plus low-level noise
@@ -68,11 +68,11 @@ TEST_CASE("FFT cross-correlation with noise still locks on target", "[correlatio
     uint32_t seed = 42;
     for (size_t i = 0; i < target.size(); ++i) {
         seed = seed * 1664525u + 1013904223u;
-        float r = (static_cast<int32_t>(seed) / 2.0e9f);  // roughly -1..1
+        float r = (static_cast<float>(static_cast<int32_t>(seed)) / 2.0e9f);  // roughly -1..1
         target[i] = 0.05f * r;
     }
     for (size_t i = 0; i < reference.size(); ++i) {
-        target[true_lag + i] += reference[i];
+        target[static_cast<std::size_t>(true_lag) + i] += reference[i];
     }
 
     auto result = mwaac::cross_correlate_fft(reference, target);
@@ -84,14 +84,14 @@ TEST_CASE("FFT correlation agrees with naive implementation", "[correlation][fft
     // Build a small case where we can run both implementations
     std::vector<float> reference(500);
     for (size_t i = 0; i < reference.size(); ++i) {
-        reference[i] = std::sin(2.0f * M_PI * 5.0f * i / 500.0f)
-                     + 0.25f * std::cos(2.0f * M_PI * 17.0f * i / 500.0f);
+        reference[i] = std::sin(2.0f * static_cast<float>(M_PI) * 5.0f * static_cast<float>(i) / 500.0f)
+                     + 0.25f * std::cos(2.0f * static_cast<float>(M_PI) * 17.0f * static_cast<float>(i) / 500.0f);
     }
 
     const int64_t true_lag = 220;
     std::vector<float> target(1500, 0.0f);
     for (size_t i = 0; i < reference.size(); ++i) {
-        target[true_lag + i] = reference[i];
+        target[static_cast<std::size_t>(true_lag) + i] = reference[i];
     }
 
     auto fft_result = mwaac::cross_correlate_fft(reference, target);
